requirements.txt
boto3>=1.35.0
opensearch-py>=2.7.0
requests-aws4auth>=1.3.0
pdfplumber>=0.11.0
python-docx>=1.1.0
tiktoken>=0.7.0
streamlit>=1.38.0
python-dotenv>=1.0.0
requests>=2.32.0
